some fixes for image builder (#7)
git -C command is relatively new, such that our build server does not have it. Not to fret, a simple and probably better solution anyway is to set the working directory for the command.

This includes some other minor fixes, where the full path was inconsistent in some functions, and the tmpdir and rsync command for building up the compat layer were run even in dryrun mode, which is not desired.