Adding a common compat layer with requirements for sciserver (#15)
This is something that should be included in all images, and likely by default without having to specify in each. Will
think about how best to do that, but for now including in the new cellxgene image.

Also switching default repository to our new fully-public repo, where these images should be served.