Increase timeout on cellxgene loading (#21)
Big files easily exceed 20 seconds even with no failures. This is not the best solution since feedback to user
is lacking, it would be easy to give up prior to the timeout... Mostly that requires fixes in cellxgene itself for
faster loading or using the other CXG based single-cell-explorer tool.